requirements.txt
huggingface_hub==0.25.2
langchain
google
langchain_google_genai
faiss-cpu